Add copy-webpack-plugin
By default Webpack only moves the /index.html file to the dist folder,
all other files named "index.html" will be ignore. This is a problem
because the CMS file is named index.html.

To fix this we use copy-webpack-plugin which is a plugin to configure
the Webpack settings to allow the index.html files to be copied.

- install cope-webpack-plugin
- add configurations to vue.config.js